back_end/requirement.txt
appdirs==1.4.3
asgiref==3.2.7
attrs==19.3.0
autopep8==1.5.2
black==19.10b0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.5
django-filter==2.2.0
django-guardian==2.2.0
djangorestframework==3.11.0
idna==2.9
itypes==1.2.0
Jinja2==2.11.2
Markdown==3.2.1
MarkupSafe==1.1.1
pathspec==0.8.0
pycodestyle==2.5.0
Pygments==2.6.1
pytz==2019.3
PyYAML==5.3.1
regex==2020.4.4
requests==2.23.0
sqlparse==0.3.1
toml==0.10.0
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.9
